bfloat16

+++ bitnet/bitnet.py
@@ -58,47 +58,6 @@ def estimate_loss(cached_batches=None):
     
     model.train()
     return out, cached_batches
-
-# class Head(nn.Module):
-#     """ one head of self-attention """
-
-#     def __init__(self, head_size):
-#         super().__init__()
-#         self.key = BitLinear(n_embd, head_size, bias=False)
-#         self.query = BitLinear(n_embd, head_size, bias=False)
-#         self.value = BitLinear(n_embd, head_size, bias=False)
-#         self.register_buffer('tril', torch.tril(torch.ones(block_size, block_size)))
-#         self.dropout = nn.Dropout(dropout)
-
-#     def forward(self, x):
-#         # input of size (batch, time-step, channels)
-#         # output of size (batch, time-step, head size)
-#         B,T,C = x.shape
-#         k = self.key(x)   # (B,T,hs)
-#         q = self.query(x) # (B,T,hs)
-#         # compute attention scores ("affinities")
-#         wei = q @ k.transpose(-2,-1) * k.shape[-1]**-0.5 # (B, T, hs) @ (B, hs, T) -> (B, T, T)
-#         wei = wei.masked_fill(self.tril[:T, :T] == 0, float('-inf')) # (B, T, T)
-#         wei = F.softmax(wei, dim=-1) # (B, T, T)
-#         wei = self.dropout(wei)
-#         # perform the weighted aggregation of the values
-#         v = self.value(x) # (B,T,hs)
-#         out = wei @ v # (B, T, T) @ (B, T, hs) -> (B, T, hs)
-#         return out
-
-# class MultiHeadAttention(nn.Module):
-#     """ multiple heads of self-attention in parallel """
-
-#     def __init__(self, num_heads, head_size):
-#         super().__init__()
-#         self.heads = nn.ModuleList([Head(head_size) for _ in range(num_heads)])
-#         self.proj = BitLinear(head_size * num_heads, n_embd)
-#         self.dropout = nn.Dropout(dropout)
-
-#     def forward(self, x):
-#         out = torch.cat([h(x) for h in self.heads], dim=-1)
-#         out = self.dropout(self.proj(out))
-#         return out
     
 class RotaryMHA(nn.Module):
     def __init__(self, d_model, n_head, n_kv_head):
@@ -261,7 +220,7 @@ if __name__ == "__main__":
     # ------------
     model = GPTLanguageModel()
     calculate_model_size_in_gb(model)
-    m = model.to(device).half()
+    m = model.to(device).bfloat16()
     if torch.cuda.is_available(): m = torch.compile(m)
     print_model_params(m)
     if torch.cuda.is_available(): training_step = torch.compile(training_step)
